requirements.txt
python-dotenv~=1.0.0
Jinja2~=3.1.2
matplotlib~=3.7.1
pillow~=9.5.0